@@ -1,12 +1,11 @@
-var isArray = (typeof Array.isArray === 'function') ?
-  // use native function
-  Array.isArray :
-  // use instanceof operator
-  function(a) {
-    return a instanceof Array;
-  };
+const isArray =
+  typeof Array.isArray === 'function'
+    ? Array.isArray
+    : function(a) {
+      return a instanceof Array;
+    };
 
-var diffFilter = function trivialMatchesDiffFilter(context) {
+export const diffFilter = function trivialMatchesDiffFilter(context) {
   if (context.left === context.right) {
     context.setResult(undefined).exit();
     return;
@@ -22,7 +21,10 @@ var diffFilter = function trivialMatchesDiffFilter(context) {
     context.setResult([context.left, 0, 0]).exit();
     return;
   }
-  if (typeof context.left === 'function' || typeof context.right === 'function') {
+  if (
+    typeof context.left === 'function' ||
+    typeof context.right === 'function'
+  ) {
     throw new Error('functions are not supported');
   }
   context.leftType = context.left === null ? 'null' : typeof context.left;
@@ -48,16 +50,17 @@ var diffFilter = function trivialMatchesDiffFilter(context) {
 
   if (context.left instanceof RegExp) {
     if (context.right instanceof RegExp) {
-      context.setResult([context.left.toString(), context.right.toString()]).exit();
+      context
+        .setResult([context.left.toString(), context.right.toString()])
+        .exit();
     } else {
       context.setResult([context.left, context.right]).exit();
-      return;
     }
   }
 };
 diffFilter.filterName = 'trivial';
 
-var patchFilter = function trivialMatchesPatchFilter(context) {
+export const patchFilter = function trivialMatchesPatchFilter(context) {
   if (typeof context.delta === 'undefined') {
     context.setResult(context.left).exit();
     return;
@@ -72,7 +75,7 @@ var patchFilter = function trivialMatchesPatchFilter(context) {
   }
   if (context.delta.length === 2) {
     if (context.left instanceof RegExp) {
-      var regexArgs = /^\/(.*)\/([gimyu]+)$/.exec(context.delta[1]);
+      const regexArgs = /^\/(.*)\/([gimyu]+)$/.exec(context.delta[1]);
       if (regexArgs) {
         context.setResult(new RegExp(regexArgs[1], regexArgs[2])).exit();
         return;
@@ -83,12 +86,11 @@ var patchFilter = function trivialMatchesPatchFilter(context) {
   }
   if (context.delta.length === 3 && context.delta[2] === 0) {
     context.setResult(undefined).exit();
-    return;
   }
 };
 patchFilter.filterName = 'trivial';
 
-var reverseFilter = function trivialReferseFilter(context) {
+export const reverseFilter = function trivialReferseFilter(context) {
   if (typeof context.delta === 'undefined') {
     context.setResult(context.delta).exit();
     return;
@@ -107,11 +109,6 @@ var reverseFilter = function trivialReferseFilter(context) {
   }
   if (context.delta.length === 3 && context.delta[2] === 0) {
     context.setResult([context.delta[0]]).exit();
-    return;
   }
 };
 reverseFilter.filterName = 'trivial';
-
-exports.diffFilter = diffFilter;
-exports.patchFilter = patchFilter;
-exports.reverseFilter = reverseFilter;
