@@ -1,60 +1,65 @@
-
-var Processor = function Processor(options){
-  this.selfOptions = options || {};
-  this.pipes = {};
-};
-
-Processor.prototype.options = function(options) {
-  if (options) {
-    this.selfOptions = options;
+class Processor {
+  constructor(options) {
+    this.selfOptions = options || {};
+    this.pipes = {};
   }
-  return this.selfOptions;
-};
 
-Processor.prototype.pipe = function(name, pipe) {
-  if (typeof name === 'string') {
-    if (typeof pipe === 'undefined') {
-      return this.pipes[name];
-    } else {
-      this.pipes[name] = pipe;
+  options(options) {
+    if (options) {
+      this.selfOptions = options;
     }
+    return this.selfOptions;
   }
-  if (name && name.name) {
-    pipe = name;
-    if (pipe.processor === this) { return pipe; }
-    this.pipes[pipe.name] = pipe;
-  }
-  pipe.processor = this;
-  return pipe;
-};
 
-Processor.prototype.process = function(input, pipe) {
-  var context = input;
-  context.options = this.options();
-  var nextPipe = pipe || input.pipe || 'default';
-  var lastPipe, lastContext;
-  while (nextPipe) {
-    if (typeof context.nextAfterChildren !== 'undefined') {
-      // children processed and coming back to parent
-      context.next = context.nextAfterChildren;
-      context.nextAfterChildren = null;
-    }
-
-    if (typeof nextPipe === 'string') {
-      nextPipe = this.pipe(nextPipe);
-    }
-    nextPipe.process(context);
-    lastContext = context;
-    lastPipe = nextPipe;
-    nextPipe = null;
-    if (context) {
-      if (context.next) {
-        context = context.next;
-        nextPipe = lastContext.nextPipe || context.pipe || lastPipe;
+  pipe(name, pipeArg) {
+    let pipe = pipeArg;
+    if (typeof name === 'string') {
+      if (typeof pipe === 'undefined') {
+        return this.pipes[name];
+      } else {
+        this.pipes[name] = pipe;
       }
     }
+    if (name && name.name) {
+      pipe = name;
+      if (pipe.processor === this) {
+        return pipe;
+      }
+      this.pipes[pipe.name] = pipe;
+    }
+    pipe.processor = this;
+    return pipe;
   }
-  return context.hasResult ? context.result : undefined;
-};
 
-exports.Processor = Processor;
+  process(input, pipe) {
+    let context = input;
+    context.options = this.options();
+    let nextPipe = pipe || input.pipe || 'default';
+    let lastPipe;
+    let lastContext;
+    while (nextPipe) {
+      if (typeof context.nextAfterChildren !== 'undefined') {
+        // children processed and coming back to parent
+        context.next = context.nextAfterChildren;
+        context.nextAfterChildren = null;
+      }
+
+      if (typeof nextPipe === 'string') {
+        nextPipe = this.pipe(nextPipe);
+      }
+      nextPipe.process(context);
+      lastContext = context;
+      lastPipe = nextPipe;
+      nextPipe = null;
+      if (context) {
+        if (context.next) {
+          context = context.next;
+          nextPipe = lastContext.nextPipe || context.pipe || lastPipe;
+        }
+      }
+    }
+    return context.hasResult ? context.result : undefined;
+  }
+}
+
+export default Processor;
